+++ changelog.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.6] Unreleased [Github repo]
+## [0.1.7] Unreleased [Github repo]
+## [0.1.6] - 2020-01-10
 ### Added
-- Reset option in "init" S_UPYDEVICE
+- Reset option in "init" S_UPYDEVICE for upydev 'sh_srepl' compatibility
 ## [0.1.5] - 2020-01-07
 ### Added
 - phantom_wr module like phantom but faster